requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==1.11.29
django-bootstrap3==11.1.0
django-bootstrap4==1.1.0
django-heroku==0.3.1
django-paypal==1.0.0
django-registration==2.4.1
django-tinymce==2.9.0
djangorestframework==3.10.3
gunicorn==19.9.0
idna==3.3
Pillow==8.4.0
psycopg2==2.8.4
python-decouple==3.6
pytz==2022.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
typing-extensions==4.1.1
urllib3==1.26.9
whitenoise==4.1
